requirements.txt
numpy==1.22.0
torch==2.0.1
torchvision==0.16.0
torchaudio==2.1.0.dev20230719
torchmetrics==1.3.2
torch-tb-profiler==0.4.3
scikit-learn==1.3.2
matplotlib==3.6.3
pandas==1.4.4
tqdm==4.65.0
requests==2.28.2
Pillow==9.4.0
